services: housekeeping: rename functions, move them around.
`ldap_register()` didn't do any registration with LDAP at all,
and was poorly named. Reflect that in the new name - a somewhat
ambiguous `register_or_update()`, and somewhat better in its
heredoc.

Move `ldap_login_func()` to the top of file. The order is now
both alphabetic and follows the call propagation.